Start of simple scene serialization
I am starting to lay down the groundwork for scene serialization and loading. The code still has a lot of room for improvement. So far, the scene is loaded from a YAML file. A lot of assumptions are made when parsing the file so I need to generalize the code so it doesn't get littered with all possible types and possibly create a generic serialization method for something like game object components since that would be tedious to manage manually.