Creating resource library and meta files
I am adding a resource manager and meta file creator to allow engine resources to be defined external to the engine. It currently uses ungisned int values for IDs. I plan to eventually move to UUIDs of some sort in the future. Meta files are stored alongside resources and currently only contain the ID of the object.