Split File/Resource serializers into separate classes
As I am developing the engine's serialization framework I have decided to separate the file and resource serializers. The reason for this is certain files on disk can contain multiple resources. For instance, a collada file can contain multiple meshes, materials, etc so there is a one to many relationship. Because of this, it will keep the code cleaner and more maintainable to separate the two.
Also, in the sample level (generatedLevel.derylevel) now loads the mesh via the mesh's uuid.