Game components now all are registered with the factory
All game components are now registering with the game component factory except for MaterialRefresher which is currently causing some compiler issues. Not all of the components are fully functional, however, because of more refactoring needed. More precisely, some components that require access to input devices will not work because of a lack of a global input manager.